+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Change CSW harvesters XML processor from lxml to saxonche [#3380](https://github.com/opendatateam/udata/pull/3380)
+- **breaking change** feat: topic elements [#3318](https://github.com/opendatateam/udata/pull/3318) [#3416](https://github.com/opendatateam/udata/pull/3416) [#3417](https://github.com/opendatateam/udata/pull/3417) [#3418](https://github.com/opendatateam/udata/pull/3418)
 
 ## 10.9.0 (2025-08-28)
 
@@ -91,7 +92,6 @@
 - Allow global admins to manually run harvesters in production [#3340](https://github.com/opendatateam/udata/pull/3340)
 - Add monthly metrics to datasets, dataservices, reuses and organizations [#3342](https://github.com/opendatateam/udata/pull/3342)
 - Fix user avatar deletion [#3346](https://github.com/opendatateam/udata/pull/3346)
-- **breaking change** feat: topic elements [#3318](https://github.com/opendatateam/udata/pull/3318)
 
 ## 10.4.2 (2025-06-05)